requirements.txt
flask
ccxt
asyncio
pandas
json
numpy
websockets
python-telegram-bot
dotenv